requirements.txt
absl-py==2.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
farama-notifications==0.0.4
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
grpcio==1.76.0
gymnasium==1.2.2
jinja2==3.1.6
kiwisolver==1.4.9
markdown==3.10
markupsafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
stable-baselines3==2.7.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.9.0
typing-extensions==4.15.0
tzdata==2025.2
werkzeug==3.1.3
hydra-core==1.3.2
